v1.1.16 Synchronize the latest code for the customized version of 4logit

+++ includes/openvpn.php
@@ -1,9 +1,7 @@
 <?php
 
-require_once 'includes/status_messages.php';
 require_once 'includes/config.php';
 require_once 'includes/wifi_functions.php';
-require_once 'app/lib/uploader.php';
 
 getWifiInterface();
 
@@ -12,7 +10,8 @@ getWifiInterface();
  */
 function DisplayOpenVPNConfig()
 {
-    $status = new StatusMessages();
+    $model = getModel();
+    $status = new \ElastPro\Messages\StatusMessage;
 
     $cipher=array('BF-CBC', 'DES-EDE-CBC', 'DES-EDE3-CBC', 'AES-128-CBC', 'AES-192-CBC',
                 'AES-256-CBC');
@@ -22,9 +21,22 @@ function DisplayOpenVPNConfig()
             'RSA-SHA512', 'SHA224', 'SHA256', 'SHA384', 'SHA512');
 
     if (isset($_POST['SaveOpenVPNSettings']) || isset($_POST['ApplyOpenVpnSettings'])) {
+        // check path
+        $clientPath = "/etc/openvpn/client";
+        if (!is_dir($clientPath)) {
+            mkdir($clientPath, 0777, true);
+        }
+
+        $serverPath = "/etc/openvpn/server";
+        if (!is_dir($serverPath)) {
+            mkdir($serverPath, 0777, true);
+        }
+
         saveOpenVpnConfig($status);
+        $status->addMessage('Configuration updated.', 'success');
         if (isset($_POST['ApplyOpenVpnSettings'])) {
-            $status->addMessage('Attempting to stop OpenVPN', 'info');
+            // $status->addMessage('Attempting to stop OpenVPN', 'info');
+            
             if (isset($_POST['role'])) { 
                 $role = $_POST['role'];
             }
@@ -33,21 +45,32 @@ function DisplayOpenVPNConfig()
                 $type = $_POST['type'];
             }
 
-            exec('sudo /bin/systemctl stop openvpn-client@client', $return);
-            exec('sudo /bin/systemctl disable openvpn-client@client', $return);
-            exec('sudo /bin/systemctl stop openvpn-server@server', $return);
-            exec('sudo /bin/systemctl disable openvpn-server@server', $return);
+            if (model_category('no_buildroot')) {
+                exec('sudo /bin/systemctl stop openvpn-client@client', $return);
+                exec('sudo /bin/systemctl disable openvpn-client@client', $return);
+                exec('sudo /bin/systemctl stop openvpn-server@server', $return);
+                exec('sudo /bin/systemctl disable openvpn-server@server', $return);
+            } else {
+                exec('sudo /etc/init.d/S60openvpn stop');
+            }
+            
             sleep(1);
             if ($type != 'off') {
-                $status->addMessage('Attempting to start OpenVPN', 'info');
-                if ($role == 'client') {
-                    exec('sudo /bin/systemctl enable openvpn-client@client', $return);
-                    exec('sudo /bin/systemctl start openvpn-client@client', $return);
-                    exec("sudo /etc/raspap/openvpn/configauth.sh $tmp_ovpn $auth_flag " .$_SESSION['ap_interface'], $return);
+                // $status->addMessage('Attempting to start OpenVPN', 'info');
+                if (model_category('no_buildroot')) {
+                    if ($role == 'client') {
+                        exec('sudo /bin/systemctl enable openvpn-client@client', $return);
+                        exec('sudo /bin/systemctl start openvpn-client@client', $return);
+                        exec("sudo /etc/raspap/openvpn/configauth.sh $tmp_ovpn $auth_flag " .$_SESSION['ap_interface'], $return);
+                    } else {
+                        exec('sudo /bin/systemctl enable openvpn-server@server', $return);
+                        exec('sudo /bin/systemctl start openvpn-server@server', $return);
+                    }
                 } else {
-                    exec('sudo /bin/systemctl enable openvpn-server@server', $return);
-                    exec('sudo /bin/systemctl start openvpn-server@server', $return);
+                    exec('sudo /etc/init.d/S60openvpn restart');
                 }
+
+                $status->addMessage('Configuration applied.', 'success');
             } else {
                 system('sudo rm /etc/openvpn/client/client.conf', $return);
                 system('sudo rm /etc/openvpn/server/server.conf', $return);
@@ -290,8 +313,13 @@ function saveConfigs($status, $role)
 
 function saveUserPass($status, $role)
 {   
+    // check checkpsw.sh
+    if (!file_exists("/etc/openvpn/server/checkpsw.sh")) {
+        exec("sudo cp /var/www/html/installers/checkpsw.sh /etc/openvpn/server/checkpsw.sh");
+    }
+
     if (strlen($_POST['text_user_pwd']) > 5) {
-        $authUserPwd = strip_tags(trim($_POST['text_user_pwd']));
+        $authUserPwd = strtr(strip_tags(trim($_POST['text_user_pwd'])), ["\r\n" => "\n"]);
     } else {
         if ($role == 'client') {
             exec("sudo rm /etc/openvpn/client/login.conf");
@@ -336,7 +364,7 @@ function SaveOpenvpnUpload($status, $file, $role)
             throw new RuntimeException('Invalid parameters');
         }
 
-        $upload = \RaspAP\Uploader\Upload::factory('vpn' . $num, $tmp_destdir);
+        $upload = \ElastPro\Uploader\FileUpload::factory('vpn' . $num, $tmp_destdir);
         $upload->set_max_file_size(64*KB);
         $upload->set_allowed_mime_types(array('text/plain', 'application/octet-stream'));
         $upload->file($file);
@@ -391,7 +419,7 @@ function SaveOVPNConfig($status, $file, $role)
             throw new RuntimeException('Invalid parameters');
         }
 
-        $upload = \RaspAP\Uploader\Upload::factory('ovpn',$tmp_destdir);
+        $upload = \ElastPro\Uploader\FileUpload::factory('ovpn',$tmp_destdir);
         $upload->set_max_file_size(64*KB);
         $upload->set_allowed_mime_types(array('ovpn' => 'text/plain'));
         $upload->file($file);
